+++ src/main/java/de/skyengine/game/world/chunk/palette/PalettedContainer.java
@@ -9,31 +9,31 @@ import java.util.Arrays;
  * wird gar kein Index-Speicher allokiert (Single-Value-Optimierung).
  *
  * <p>bitsPerEntry wächst automatisch mit der Palette (max. 15 Bit bei 32³ Zellen). API in
- * {@code short} State-IDs — kompatibel zur bestehenden Chunk-Schicht.
+ * {@code int} State-IDs — die Palette entkoppelt die Speichergröße von der ID-Breite.
  */
 public final class PalettedContainer {
 
     private final int size;
-    private short[] palette;
+    private int[] palette;
     private int paletteSize;
     private BitStorage storage;   // null => Single-Value (palette[0])
     private int nonAir;
 
-    public PalettedContainer(int size, short fill) {
+    public PalettedContainer(int size, int fill) {
         this.size = size;
-        this.palette = new short[4];
+        this.palette = new int[4];
         this.palette[0] = fill;
         this.paletteSize = 1;
         this.nonAir = fill == 0 ? 0 : size;
     }
 
-    public short get(int index) {
+    public int get(int index) {
         if (this.storage == null) return this.palette[0];
         return this.palette[this.storage.get(index)];
     }
 
-    public void set(int index, short stateId) {
-        short old = get(index);
+    public void set(int index, int stateId) {
+        int old = get(index);
         if (old == stateId) return;
 
         int id = idFor(stateId);
@@ -49,7 +49,7 @@ public final class PalettedContainer {
 
     /* Liefert (oder vergibt) den Paletten-Index einer State-ID; vergrößert Palette/Storage.
        Linearer Scan über die (typisch winzige) Palette - allokationsfrei und ohne Boxing. */
-    private int idFor(short stateId) {
+    private int idFor(int stateId) {
         for (int i = 0; i < this.paletteSize; i++) {
             if (this.palette[i] == stateId) return i;
         }
@@ -79,7 +79,7 @@ public final class PalettedContainer {
 
     /* --- Zugriff für Persistenz (Phase: Chunk-Save) --- */
 
-    public short[] paletteEntries() {
+    public int[] paletteEntries() {
         return Arrays.copyOf(this.palette, this.paletteSize);
     }
